Extract UpcomingEventCard, EventEditForm, CreateEventForm, events-shared helpers
Lifts the events-route UI into shared components so dashboard._index.tsx can
render the same upcoming-event cards (with inline RSVP, member-response columns,
edit-in-place) without duplicating ~360 lines of JSX.

- components/UpcomingEventCard.tsx — the upcoming-event card
- components/EventEditForm.tsx — extracted from inline function in events route
- components/CreateEventForm.tsx — extracted create form
- lib/events-shared.ts — client-safe types + helpers (EventCard, sortEventsBySchedule, etc.)

The events route now imports these instead of inlining the JSX. Form _action
values are renamed to event_create / event_update / event_rsvp so the same
components work from either route's action handler. Existing tests updated
to the new action names; all 617 tests pass.

Prep for #145 (events migration onto the dashboard).
