+++ app/app/routes/dashboard.events.tsx
@@ -1,4 +1,4 @@
-import { Form, Link, isRouteErrorResponse, useNavigation } from "react-router";
+import { Link, isRouteErrorResponse, useNavigation } from "react-router";
 import { useEffect, useRef, useState } from "react";
 import type { Route } from "./+types/dashboard.events";
 import { requireActiveUser } from "../lib/auth.server";
@@ -8,206 +8,28 @@ import {
   runRsvpEventAction,
   runUpdateEventAction,
 } from "../lib/event-actions.server";
-import { EventRestaurantFields } from "../components/EventRestaurantFields";
-import { formatDateForDisplay, formatTimeForDisplay, getAppTimeZone, isEventInPastInTimeZone } from "../lib/dateUtils";
-import type { Event, RsvpWithUser } from "../lib/types";
-import { Alert, Badge, Button, Card, EmptyState, PageHeader, UserAvatar } from "../components/ui";
-import { CalendarDaysIcon, CheckIcon, ChevronDownIcon, ClockIcon, MapPinIcon, PencilSquareIcon, XMarkIcon } from "@heroicons/react/24/outline";
-
-interface EventRow extends Event {
-  creator_name: string | null;
-  creator_email: string | null;
-}
-
-interface MemberRow {
-  id: number;
-  name: string | null;
-  email: string;
-  picture: string | null;
-}
-
-interface EventCard extends EventRow {
-  canEdit: boolean;
-  creatorLabel: string;
-  displayStatus?: "completed" | "cancelled";
-  userRsvp?: { status: string; comments: string | null } | null;
-  allRsvps?: RsvpWithUser[];
-  notResponded?: MemberRow[];
-}
-
-interface EventFormState {
-  restaurantName: string;
-  restaurantAddress: string;
-  eventDate: string;
-  eventTime: string;
-}
-
-type EventMutationAction = "create" | "update" | "rsvp";
-
-const EMPTY_EVENT_FORM: EventFormState = {
-  restaurantName: "",
-  restaurantAddress: "",
-  eventDate: "",
-  eventTime: "18:00",
-};
-
-function getEventDateTimeKey(event: Pick<Event, "event_date" | "event_time">): string {
-  return `${event.event_date}T${event.event_time || "18:00"}`;
-}
-
-function sortEventsBySchedule<T extends Pick<Event, "event_date" | "event_time">>(
-  events: T[],
-  direction: "asc" | "desc"
-): T[] {
-  return [...events].sort((left, right) => {
-    const comparison = getEventDateTimeKey(left).localeCompare(getEventDateTimeKey(right));
-    return direction === "asc" ? comparison : -comparison;
-  });
-}
-
-function getEventResponseCounts(event: Pick<EventCard, "allRsvps" | "notResponded">) {
-  const allRsvps = event.allRsvps || [];
-
-  return {
-    yes: allRsvps.filter((rsvp) => rsvp.status === "yes").length,
-    maybe: allRsvps.filter((rsvp) => rsvp.status === "maybe").length,
-    no: allRsvps.filter((rsvp) => rsvp.status === "no").length,
-    pending: event.notResponded?.length || 0,
-  };
-}
-
-function getUserRsvpBadge(status: string | null | undefined) {
-  if (status === "yes") {
-    return { label: "You're in", variant: "success" as const };
-  }
-
-  if (status === "maybe") {
-    return { label: "You're maybe", variant: "warning" as const };
-  }
-
-  if (status === "no") {
-    return { label: "You're out", variant: "danger" as const };
-  }
-
-  return { label: "RSVP needed", variant: "muted" as const };
-}
-
-function getCreatorLabel(event: EventRow, currentUserId: number): string {
-  if (event.created_by === currentUserId) {
-    return "Created by you";
-  }
-
-  if (event.creator_name) {
-    return `Created by ${event.creator_name}`;
-  }
-
-  if (event.creator_email) {
-    return `Created by ${event.creator_email}`;
-  }
-
-  return "Created by an admin";
-}
-
-function toEditableState(event: EventCard): EventFormState & { id: number } {
-  return {
-    id: event.id,
-    restaurantName: event.restaurant_name,
-    restaurantAddress: event.restaurant_address || "",
-    eventDate: event.event_date,
-    eventTime: event.event_time || "18:00",
-  };
-}
-
-interface EventEditFormProps {
-  event: EventCard;
-  formData: { id: number } & EventFormState;
-  idPrefix: string;
-  onCancel: () => void;
-  onRestaurantNameChange: (value: string) => void;
-  onRestaurantAddressChange: (value: string) => void;
-  onEventDateChange: (value: string) => void;
-  onEventTimeChange: (value: string) => void;
-}
-
-function EventEditForm({
-  event,
-  formData,
-  idPrefix,
-  onCancel,
-  onRestaurantNameChange,
-  onRestaurantAddressChange,
-  onEventDateChange,
-  onEventTimeChange,
-}: EventEditFormProps) {
-  return (
-    <Form method="post" preventScrollReset className="space-y-4">
-      <input type="hidden" name="_action" value="update" />
-      <input type="hidden" name="id" value={formData.id} />
-
-      <div className="flex items-start justify-between gap-4">
-        <div>
-          <h3 className="text-xl font-semibold text-foreground">Edit Event</h3>
-          <p className="mt-1 text-sm text-muted-foreground">{event.creatorLabel}</p>
-        </div>
-        <Button type="button" variant="secondary" size="sm" onClick={onCancel}>
-          Cancel
-        </Button>
-      </div>
-
-      <EventRestaurantFields
-        restaurantName={formData.restaurantName}
-        restaurantAddress={formData.restaurantAddress}
-        onRestaurantNameChange={onRestaurantNameChange}
-        onRestaurantAddressChange={onRestaurantAddressChange}
-      />
-
-      <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-        <div>
-          <label htmlFor={`${idPrefix}-date-${event.id}`} className="mb-1 block text-sm font-medium text-foreground">
-            Event Date *
-          </label>
-          <input
-            id={`${idPrefix}-date-${event.id}`}
-            name="event_date"
-            type="date"
-            required
-            value={formData.eventDate}
-            onChange={(currentEvent) => onEventDateChange(currentEvent.target.value)}
-            className="w-full rounded-md border border-border px-3 py-2 focus:outline-none focus:ring-2 focus:ring-accent"
-          />
-        </div>
-
-        <div>
-          <label htmlFor={`${idPrefix}-time-${event.id}`} className="mb-1 block text-sm font-medium text-foreground">
-            Event Time
-          </label>
-          <input
-            id={`${idPrefix}-time-${event.id}`}
-            name="event_time"
-            type="time"
-            value={formData.eventTime}
-            onChange={(currentEvent) => onEventTimeChange(currentEvent.target.value)}
-            className="w-full rounded-md border border-border px-3 py-2 focus:outline-none focus:ring-2 focus:ring-accent"
-          />
-        </div>
-      </div>
-
-      <label className="flex items-center gap-2 text-sm text-foreground">
-        <input
-          id={`${idPrefix}-send-updates-${event.id}`}
-          name="send_updates"
-          type="checkbox"
-          value="true"
-          defaultChecked={true}
-          className="h-4 w-4 rounded border-border text-accent focus:ring-accent"
-        />
-        Send calendar updates to all active members
-      </label>
-
-      <Button type="submit">Save Changes</Button>
-    </Form>
-  );
-}
+import {
+  EMPTY_EVENT_FORM,
+  getCreatorLabel,
+  sortEventsBySchedule,
+  toEditableState,
+  type EventCard,
+  type EventFormState,
+  type EventMemberRow,
+  type EventRow,
+} from "../lib/events-shared";
+import { CreateEventForm } from "../components/CreateEventForm";
+import { EventEditForm } from "../components/EventEditForm";
+import { UpcomingEventCard } from "../components/UpcomingEventCard";
+import {
+  formatDateForDisplay,
+  formatTimeForDisplay,
+  getAppTimeZone,
+  isEventInPastInTimeZone,
+} from "../lib/dateUtils";
+import type { RsvpWithUser } from "../lib/types";
+import { Alert, Badge, Button, Card, EmptyState, PageHeader } from "../components/ui";
+import { CalendarDaysIcon, MapPinIcon, PencilSquareIcon } from "@heroicons/react/24/outline";
 
 export async function loader({ request, context }: Route.LoaderArgs) {
   const user = await requireActiveUser(request, context);
@@ -232,7 +54,7 @@ export async function loader({ request, context }: Route.LoaderArgs) {
   ]);
 
   const events = (eventsResult.results || []) as unknown as EventRow[];
-  const allMembers = (allMembersResult.results || []) as unknown as MemberRow[];
+  const allMembers = (allMembersResult.results || []) as unknown as EventMemberRow[];
   const appTimeZone = getAppTimeZone(context.cloudflare.env.APP_TIMEZONE);
 
   const upcomingEventsRaw = sortEventsBySchedule(
@@ -242,12 +64,11 @@ export async function loader({ request, context }: Route.LoaderArgs) {
         !isEventInPastInTimeZone(event.event_date, event.event_time || "18:00", appTimeZone)
     ),
     "asc"
-  )
-    .map((event) => ({
-      ...event,
-      canEdit: canEditEvent(user, event),
-      creatorLabel: getCreatorLabel(event, user.id),
-    }));
+  ).map((event) => ({
+    ...event,
+    canEdit: canEditEvent(user, event),
+    creatorLabel: getCreatorLabel(event, user.id),
+  }));
 
   const pastEvents = sortEventsBySchedule(
     events.filter(
@@ -257,11 +78,11 @@ export async function loader({ request, context }: Route.LoaderArgs) {
     ),
     "desc"
   ).map((event) => ({
-      ...event,
-      canEdit: canEditEvent(user, event),
-      creatorLabel: getCreatorLabel(event, user.id),
-      displayStatus: event.status === "cancelled" ? "cancelled" : "completed",
-    })) as EventCard[];
+    ...event,
+    canEdit: canEditEvent(user, event),
+    creatorLabel: getCreatorLabel(event, user.id),
+    displayStatus: event.status === "cancelled" ? "cancelled" : "completed",
+  })) as EventCard[];
 
   const upcomingEvents = (await Promise.all(
     upcomingEventsRaw.map(async (event) => {
@@ -309,7 +130,7 @@ export async function action({ request, context }: Route.ActionArgs) {
   const user = await requireActiveUser(request, context);
   const db = context.cloudflare.env.DB;
   const formData = await request.formData();
-  const actionType = String(formData.get("_action") || "rsvp");
+  const actionType = String(formData.get("_action") || "event_rsvp");
   const ctx = {
     db,
     queue: context.cloudflare.env.EMAIL_DELIVERY_QUEUE,
@@ -319,11 +140,11 @@ export async function action({ request, context }: Route.ActionArgs) {
     route: "/dashboard/events",
   };
 
-  if (actionType === "create") {
+  if (actionType === "event_create") {
     return runCreateEventAction(ctx);
   }
 
-  if (actionType === "update") {
+  if (actionType === "event_update") {
     return runUpdateEventAction(ctx);
   }
 
@@ -372,13 +193,17 @@ export default function EventsPage({ loaderData, actionData }: Route.ComponentPr
     setExpandedEventId((current) => (current === eventId ? null : eventId));
   }
 
+  function handleEditFieldChange(field: keyof EventFormState, value: string) {
+    setEditData((current) => ({ ...current, [field]: value }));
+  }
+
   useEffect(() => {
     if (navigation.state !== "submitting" || !navigation.formData) {
       return;
     }
 
     const submittedAction = navigation.formData.get("_action");
-    if (submittedAction === "create" || submittedAction === "update") {
+    if (submittedAction === "event_create" || submittedAction === "event_update") {
       submittedActionRef.current = String(submittedAction);
     }
   }, [navigation.formData, navigation.state]);
@@ -393,12 +218,12 @@ export default function EventsPage({ loaderData, actionData }: Route.ComponentPr
       return;
     }
 
-    if (submittedActionRef.current === "create") {
+    if (submittedActionRef.current === "event_create") {
       setShowCreateForm(false);
       resetCreateForm();
     }
 
-    if (submittedActionRef.current === "update") {
+    if (submittedActionRef.current === "event_update") {
       cancelEditing();
     }
 
@@ -448,75 +273,7 @@ export default function EventsPage({ loaderData, actionData }: Route.ComponentPr
       ) : null}
 
       {showCreateForm ? (
-        <Card className="mb-8 p-6">
-          <h2 className="text-xl font-semibold text-foreground">Create Ad Hoc Event</h2>
-          <p className="mt-1 text-sm text-muted-foreground">
-            Pick a restaurant from Google Places, choose the date, and skip the voting flow entirely.
-          </p>
-          <Form method="post" preventScrollReset className="mt-5 space-y-4">
-            <input type="hidden" name="_action" value="create" />
-
-            <EventRestaurantFields
-              restaurantName={createData.restaurantName}
-              restaurantAddress={createData.restaurantAddress}
-              onRestaurantNameChange={(value) =>
-                setCreateData((current) => ({ ...current, restaurantName: value }))
-              }
-              onRestaurantAddressChange={(value) =>
-                setCreateData((current) => ({ ...current, restaurantAddress: value }))
-              }
-            />
-
-            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-              <div>
-                <label htmlFor="create-event-date" className="block text-sm font-medium text-foreground mb-1">
-                  Event Date *
-                </label>
-                <input
-                  id="create-event-date"
-                  name="event_date"
-                  type="date"
-                  required
-                  value={createData.eventDate}
-                  onChange={(event) =>
-                    setCreateData((current) => ({ ...current, eventDate: event.target.value }))
-                  }
-                  className="w-full rounded-md border border-border px-3 py-2 focus:outline-none focus:ring-2 focus:ring-accent"
-                />
-              </div>
-
-              <div>
-                <label htmlFor="create-event-time" className="block text-sm font-medium text-foreground mb-1">
-                  Event Time
-                </label>
-                <input
-                  id="create-event-time"
-                  name="event_time"
-                  type="time"
-                  value={createData.eventTime}
-                  onChange={(event) =>
-                    setCreateData((current) => ({ ...current, eventTime: event.target.value }))
-                  }
-                  className="w-full rounded-md border border-border px-3 py-2 focus:outline-none focus:ring-2 focus:ring-accent"
-                />
-              </div>
-            </div>
-
-            <label className="flex items-center gap-2 text-sm text-foreground">
-              <input
-                id="create-send-invites"
-                name="send_invites"
-                type="checkbox"
-                value="true"
-                defaultChecked={true}
-                className="h-4 w-4 rounded border-border text-accent focus:ring-accent"
-              />
-              Send calendar invites to all active members
-            </label>
-
-            <Button type="submit">Create Event</Button>
-          </Form>
-        </Card>
+        <CreateEventForm formData={createData} onChange={setCreateData} />
       ) : null}
 
       <div className="mb-12">
@@ -561,364 +318,23 @@ export default function EventsPage({ loaderData, actionData }: Route.ComponentPr
             </Card>
 
             <div className="grid gap-5 xl:auto-rows-fr xl:grid-cols-2">
-            {upcomingEvents.map((event) => {
-                const counts = getEventResponseCounts(event);
-                const userRsvpBadge = getUserRsvpBadge(event.userRsvp?.status);
-                const goingRsvps = event.allRsvps?.filter((rsvp) => rsvp.status === "yes") || [];
-                const maybeRsvps = event.allRsvps?.filter((rsvp) => rsvp.status === "maybe") || [];
-                const declinedRsvps = event.allRsvps?.filter((rsvp) => rsvp.status === "no") || [];
+              {upcomingEvents.map((event) => {
                 const isExpanded = expandedEventId === event.id || editingId === event.id;
                 const isEditing = editingId === event.id;
-                const detailsId = `event-details-${event.id}`;
-                const titleId = `event-title-${event.id}`;
-
                 return (
-                  <Card
+                  <UpcomingEventCard
                     key={event.id}
-                    hover
-                    role="article"
-                    aria-labelledby={titleId}
-                    className={`overflow-hidden p-0 ${isExpanded ? "xl:col-span-2" : "h-full"}`}
-                  >
-                    <div className={`flex flex-col bg-accent/[0.04] p-5 sm:p-6 ${isExpanded ? "" : "h-full"}`}>
-                      <div className="flex flex-col gap-5 lg:flex-row lg:items-start lg:justify-between">
-                        <div className="flex flex-col gap-4 sm:flex-row sm:items-start">
-                          <div className="min-w-24 rounded-2xl border border-border/70 bg-background/80 px-4 py-3 text-center shadow-sm">
-                            <p className="text-[0.65rem] font-semibold uppercase tracking-[0.28em] text-muted-foreground">
-                              {formatDateForDisplay(event.event_date, { month: "short" })}
-                            </p>
-                            <p className="mt-1 text-3xl font-semibold leading-none text-foreground">
-                              {formatDateForDisplay(event.event_date, { day: "numeric" })}
-                            </p>
-                            <p className="mt-1 text-xs font-medium text-muted-foreground">
-                              {formatDateForDisplay(event.event_date, { weekday: "short" })}
-                            </p>
-                          </div>
-
-                          <div className="space-y-3">
-                            <div className="flex flex-wrap items-center gap-2">
-                              <Badge variant="success">Upcoming</Badge>
-                              <Badge variant={userRsvpBadge.variant}>{userRsvpBadge.label}</Badge>
-                            </div>
-
-                            <div>
-                              <h3 id={titleId} className="text-2xl font-semibold text-foreground">
-                                {event.restaurant_name}
-                              </h3>
-                              {event.restaurant_address ? (
-                                <p className="mt-2 flex items-center gap-1.5 text-sm text-muted-foreground">
-                                  <MapPinIcon className="h-4 w-4" />
-                                  {event.restaurant_address}
-                                </p>
-                              ) : null}
-                            </div>
-
-                            <div className="flex flex-wrap gap-4 text-sm text-muted-foreground">
-                              <span className="inline-flex items-center gap-1.5">
-                                <CalendarDaysIcon className="h-4 w-4" />
-                                {formatDateForDisplay(event.event_date, {
-                                  weekday: "long",
-                                  month: "long",
-                                  day: "numeric",
-                                  year: "numeric",
-                                })}
-                              </span>
-                              <span className="inline-flex items-center gap-1.5">
-                                <ClockIcon className="h-4 w-4" />
-                                {formatTimeForDisplay(event.event_time || "18:00")}
-                              </span>
-                            </div>
-
-                            <p className="text-xs uppercase tracking-[0.18em] text-muted-foreground">
-                              {event.creatorLabel}
-                            </p>
-                          </div>
-                        </div>
-
-                        <div className="flex flex-wrap items-center gap-2">
-                          {event.canEdit ? (
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="sm"
-                              aria-label={`Edit ${event.restaurant_name}`}
-                              onClick={() => startEditing(event)}
-                            >
-                              <PencilSquareIcon className="mr-1 h-4 w-4" />
-                              Edit
-                            </Button>
-                          ) : null}
-                          {!isEditing ? (
-                            <Button
-                              type="button"
-                              variant="secondary"
-                              size="sm"
-                              aria-controls={detailsId}
-                              aria-expanded={isExpanded}
-                              aria-label={`${isExpanded ? "Hide" : "Open"} details for ${event.restaurant_name}`}
-                              onClick={() => toggleUpcomingEvent(event.id)}
-                            >
-                              {isExpanded ? "Hide details" : "Open details"}
-                              <ChevronDownIcon
-                                className={`ml-1 h-4 w-4 transition-transform ${isExpanded ? "rotate-180" : ""}`}
-                              />
-                            </Button>
-                          ) : null}
-                        </div>
-                      </div>
-
-                      <div className="mt-auto grid grid-cols-2 gap-3 pt-5 xl:grid-cols-4">
-                        <div className="rounded-2xl border border-border/70 bg-background/70 p-3 shadow-sm">
-                          <p className="text-[0.65rem] font-semibold uppercase tracking-[0.22em] text-muted-foreground">
-                            Going
-                          </p>
-                          <p className="mt-2 text-2xl font-semibold text-foreground">{counts.yes}</p>
-                        </div>
-                        <div className="rounded-2xl border border-border/70 bg-background/70 p-3 shadow-sm">
-                          <p className="text-[0.65rem] font-semibold uppercase tracking-[0.22em] text-muted-foreground">
-                            Maybe
-                          </p>
-                          <p className="mt-2 text-2xl font-semibold text-foreground">{counts.maybe}</p>
-                        </div>
-                        <div className="rounded-2xl border border-border/70 bg-background/70 p-3 shadow-sm">
-                          <p className="text-[0.65rem] font-semibold uppercase tracking-[0.22em] text-muted-foreground">
-                            Out
-                          </p>
-                          <p className="mt-2 text-2xl font-semibold text-foreground">{counts.no}</p>
-                        </div>
-                        <div className="rounded-2xl border border-border/70 bg-background/70 p-3 shadow-sm">
-                          <p className="text-[0.65rem] font-semibold uppercase tracking-[0.22em] text-muted-foreground">
-                            Pending
-                          </p>
-                          <p className="mt-2 text-2xl font-semibold text-foreground">{counts.pending}</p>
-                        </div>
-                      </div>
-                    </div>
-
-                    {isExpanded ? (
-                      <div id={detailsId} className="border-t border-border/60 bg-card/80 p-5 sm:p-6">
-                        {isEditing ? (
-                          <EventEditForm
-                            event={event}
-                            formData={editData}
-                            idPrefix="edit-event"
-                            onCancel={cancelEditing}
-                            onRestaurantNameChange={(value) =>
-                              setEditData((current) => ({ ...current, restaurantName: value }))
-                            }
-                            onRestaurantAddressChange={(value) =>
-                              setEditData((current) => ({ ...current, restaurantAddress: value }))
-                            }
-                            onEventDateChange={(value) =>
-                              setEditData((current) => ({ ...current, eventDate: value }))
-                            }
-                            onEventTimeChange={(value) =>
-                              setEditData((current) => ({ ...current, eventTime: value }))
-                            }
-                          />
-                        ) : (
-                          <div className="grid gap-6 2xl:grid-cols-[minmax(0,0.95fr)_minmax(0,1.05fr)]">
-                            <div className="rounded-2xl border border-border/70 bg-muted/30 p-4 sm:p-5">
-                              <div className="flex flex-wrap items-center justify-between gap-3">
-                                <div>
-                                  <h4 className="font-semibold text-foreground">Your RSVP</h4>
-                                  <p className="mt-1 text-sm text-muted-foreground">
-                                    Update your response here without losing your place on the page.
-                                  </p>
-                                </div>
-                                <Badge variant={userRsvpBadge.variant}>{userRsvpBadge.label}</Badge>
-                              </div>
-
-                              <Form method="post" preventScrollReset className="mt-5 space-y-4">
-                                <input type="hidden" name="_action" value="rsvp" />
-                                <input type="hidden" name="event_id" value={event.id} />
-
-                                <div>
-                                  <label className="mb-2 block text-sm font-medium text-foreground">
-                                    Will you attend?
-                                  </label>
-                                  <div className="flex flex-wrap gap-3">
-                                    {["yes", "no", "maybe"].map((option) => (
-                                      <label
-                                        key={option}
-                                        className={`cursor-pointer select-none rounded-md px-4 py-2 font-medium transition-all ${
-                                          event.userRsvp?.status === option
-                                            ? "bg-accent text-background shadow-sm"
-                                            : "border border-border bg-card text-foreground hover:border-accent/50 hover:bg-muted active:scale-95 active:shadow-inner"
-                                        }`}
-                                      >
-                                        <input
-                                          type="radio"
-                                          name="status"
-                                          value={option}
-                                          defaultChecked={event.userRsvp?.status === option}
-                                          className="sr-only"
-                                          onChange={(currentEvent) => {
-                                            if (currentEvent.target.checked) {
-                                              currentEvent.target.form?.requestSubmit();
-                                            }
-                                          }}
-                                        />
-                                        {option.charAt(0).toUpperCase() + option.slice(1)}
-                                      </label>
-                                    ))}
-                                  </div>
-                                  <p className="mt-2 text-xs text-muted-foreground">
-                                    Click a response to save it immediately.
-                                  </p>
-                                </div>
-
-                                <div>
-                                  <label htmlFor={`comments-${event.id}`} className="mb-2 block text-sm font-medium text-foreground">
-                                    Comments (Optional)
-                                  </label>
-                                  <div className="space-y-2">
-                                    <textarea
-                                      id={`comments-${event.id}`}
-                                      name="comments"
-                                      defaultValue={event.userRsvp?.comments || ""}
-                                      placeholder="Any notes about your attendance"
-                                      rows={3}
-                                      className="w-full rounded-md border border-border bg-card px-3 py-2 text-foreground focus:outline-none focus:ring-2 focus:ring-accent"
-                                    />
-                                    <Button type="submit" size="sm">
-                                      Update Comments
-                                    </Button>
-                                  </div>
-                                </div>
-                              </Form>
-                            </div>
-
-                            <div className="space-y-4">
-                              <div>
-                                <h4 className="font-semibold text-foreground">Member responses</h4>
-                                <p className="mt-1 text-sm text-muted-foreground">
-                                  {counts.yes} going, {counts.maybe} maybe, {counts.no} out, {counts.pending} pending.
-                                </p>
-                              </div>
-
-                              <div className="grid gap-4 xl:grid-cols-2">
-                                {goingRsvps.length > 0 ? (
-                                  <div className="rounded-2xl border border-border/70 bg-background/70 p-4">
-                                    <div className="mb-3 flex items-center justify-between gap-3">
-                                      <Badge variant="success" className="inline-flex items-center gap-1">
-                                        <CheckIcon className="h-4 w-4" />
-                                        Going
-                                      </Badge>
-                                      <span className="text-sm text-muted-foreground">{goingRsvps.length}</span>
-                                    </div>
-                                    <div className="max-h-64 space-y-2 overflow-y-auto pr-1">
-                                      {goingRsvps.map((rsvp) => (
-                                        <div
-                                          key={rsvp.id}
-                                          className="flex items-center gap-3 rounded-md border border-border/80 bg-muted/40 p-3"
-                                        >
-                                          <UserAvatar src={rsvp.picture} name={rsvp.name} email={rsvp.email} />
-                                          <div className="flex-1">
-                                            <p className="font-medium text-foreground">{rsvp.name || rsvp.email}</p>
-                                            {rsvp.comments ? (
-                                              <p className="text-sm text-muted-foreground">{rsvp.comments}</p>
-                                            ) : null}
-                                          </div>
-                                        </div>
-                                      ))}
-                                    </div>
-                                  </div>
-                                ) : null}
-
-                                {maybeRsvps.length > 0 ? (
-                                  <div className="rounded-2xl border border-border/70 bg-background/70 p-4">
-                                    <div className="mb-3 flex items-center justify-between gap-3">
-                                      <Badge variant="warning" className="inline-flex items-center gap-1">
-                                        ? Maybe
-                                      </Badge>
-                                      <span className="text-sm text-muted-foreground">{maybeRsvps.length}</span>
-                                    </div>
-                                    <div className="max-h-64 space-y-2 overflow-y-auto pr-1">
-                                      {maybeRsvps.map((rsvp) => (
-                                        <div
-                                          key={rsvp.id}
-                                          className="flex items-center gap-3 rounded-md border border-border/70 bg-muted/30 p-3"
-                                        >
-                                          <UserAvatar src={rsvp.picture} name={rsvp.name} email={rsvp.email} />
-                                          <div className="flex-1">
-                                            <p className="font-medium text-foreground">{rsvp.name || rsvp.email}</p>
-                                            {rsvp.comments ? (
-                                              <p className="text-sm text-muted-foreground">{rsvp.comments}</p>
-                                            ) : null}
-                                          </div>
-                                        </div>
-                                      ))}
-                                    </div>
-                                  </div>
-                                ) : null}
-
-                                {declinedRsvps.length > 0 ? (
-                                  <div className="rounded-2xl border border-border/70 bg-background/70 p-4">
-                                    <div className="mb-3 flex items-center justify-between gap-3">
-                                      <Badge variant="danger" className="inline-flex items-center gap-1">
-                                        <XMarkIcon className="h-4 w-4" />
-                                        Not Going
-                                      </Badge>
-                                      <span className="text-sm text-muted-foreground">{declinedRsvps.length}</span>
-                                    </div>
-                                    <div className="max-h-64 space-y-2 overflow-y-auto pr-1">
-                                      {declinedRsvps.map((rsvp) => (
-                                        <div
-                                          key={rsvp.id}
-                                          className="flex items-center gap-3 rounded-md border border-border/60 bg-muted/20 p-3 opacity-80"
-                                        >
-                                          <UserAvatar src={rsvp.picture} name={rsvp.name} email={rsvp.email} />
-                                          <div className="flex-1">
-                                            <p className="font-medium text-foreground">{rsvp.name || rsvp.email}</p>
-                                            {rsvp.comments ? (
-                                              <p className="text-sm text-muted-foreground">{rsvp.comments}</p>
-                                            ) : null}
-                                          </div>
-                                        </div>
-                                      ))}
-                                    </div>
-                                  </div>
-                                ) : null}
-
-                                {event.notResponded && event.notResponded.length > 0 ? (
-                                  <div className="rounded-2xl border border-border/70 bg-background/70 p-4">
-                                    <div className="mb-3 flex items-center justify-between gap-3">
-                                      <div className="flex items-center gap-1.5 text-sm font-semibold text-muted-foreground">
-                                        <ClockIcon className="h-4 w-4" />
-                                        No Response Yet
-                                      </div>
-                                      <span className="text-sm text-muted-foreground">{event.notResponded.length}</span>
-                                    </div>
-                                    <div className="max-h-64 space-y-2 overflow-y-auto pr-1">
-                                      {event.notResponded.map((member) => (
-                                        <div
-                                          key={member.id}
-                                          className="flex items-center gap-3 rounded-md border border-border bg-muted/50 p-3 opacity-60"
-                                        >
-                                          <UserAvatar
-                                            src={member.picture}
-                                            name={member.name}
-                                            email={member.email}
-                                            className="grayscale"
-                                          />
-                                          <p className="font-medium text-muted-foreground">
-                                            {member.name || member.email}
-                                          </p>
-                                        </div>
-                                      ))}
-                                    </div>
-                                  </div>
-                                ) : null}
-                              </div>
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    ) : null}
-                  </Card>
+                    event={event}
+                    isExpanded={isExpanded}
+                    isEditing={isEditing}
+                    editFormData={editData}
+                    onStartEdit={() => startEditing(event)}
+                    onCancelEdit={cancelEditing}
+                    onToggleExpand={() => toggleUpcomingEvent(event.id)}
+                    onEditFieldChange={handleEditFieldChange}
+                  />
                 );
-            })}
+              })}
             </div>
           </div>
         )}
@@ -943,24 +359,21 @@ export default function EventsPage({ loaderData, actionData }: Route.ComponentPr
                     formData={editData}
                     idPrefix="past-edit-event"
                     onCancel={cancelEditing}
-                    onRestaurantNameChange={(value) =>
-                      setEditData((current) => ({ ...current, restaurantName: value }))
-                    }
+                    onRestaurantNameChange={(value) => handleEditFieldChange("restaurantName", value)}
                     onRestaurantAddressChange={(value) =>
-                      setEditData((current) => ({ ...current, restaurantAddress: value }))
+                      handleEditFieldChange("restaurantAddress", value)
                     }
-                    onEventDateChange={(value) =>
-                      setEditData((current) => ({ ...current, eventDate: value }))
-                    }
-                    onEventTimeChange={(value) =>
-                      setEditData((current) => ({ ...current, eventTime: value }))
-                    }
+                    onEventDateChange={(value) => handleEditFieldChange("eventDate", value)}
+                    onEventTimeChange={(value) => handleEditFieldChange("eventTime", value)}
                   />
                 ) : (
                   <div className="flex flex-col gap-4 sm:flex-row sm:items-start sm:justify-between">
                     <div className="flex-1">
                       <div className="mb-3 flex flex-wrap items-center gap-3">
-                        <h3 id={`past-event-title-${event.id}`} className="text-xl font-semibold text-foreground">
+                        <h3
+                          id={`past-event-title-${event.id}`}
+                          className="text-xl font-semibold text-foreground"
+                        >
                           {event.restaurant_name}
                         </h3>
                         <Badge variant={event.displayStatus === "completed" ? "muted" : "danger"}>
@@ -1004,10 +417,7 @@ export default function EventsPage({ loaderData, actionData }: Route.ComponentPr
 export function HydrateFallback() {
   return (
     <main className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
-      <PageHeader
-        title="Events"
-        description="Upcoming and past Meatup.Club events"
-      />
+      <PageHeader title="Events" description="Upcoming and past Meatup.Club events" />
       <div className="space-y-6">
         <Card className="animate-pulse bg-muted/30 p-6">
           <div className="mb-4 h-6 w-56 rounded bg-muted" />
